+++ src/main/java/codeemoji/inlay/vcs/RefactorService.java
@@ -42,7 +42,7 @@ import java.util.function.Function;
 
 @Service(Level.PROJECT)
 public final class RefactorService implements Disposable {
-
+//TODO: fix ... args not creating matching method signatures
     private static final Logger LOG = Logger.getInstance(RefactorService.class);
     private static final int SCAN_DELAY_MS = 5;
     private static final int DEFAULT_MAX_COMMITS = 10;
